requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
pyaml==15.8.2
django-toolbelt==0.0.1
django-extensions==1.6.1
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
cryptography==0.9.3
django-tools==0.29.5
django-crispy-forms==1.5.2
requests==2.9.0
responses==0.5.0
